requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
deep-translator==1.11.4
fire==0.5.0
idna==3.6
pyaml==23.12.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
termcolor==2.4.0
tqdm==4.66.2
urllib3==2.2.1
